Move trainer runner local test to non gpu runner
