+++ rllib/core/rl_trainer/tests/test_trainer_runner.py
@@ -8,10 +8,23 @@ from ray.rllib.core.rl_trainer.trainer_runner import TrainerRunner
 from ray.rllib.core.testing.tf.bc_module import DiscreteBCTFModule
 from ray.rllib.core.testing.tf.bc_rl_trainer import BCTfRLTrainer
 from ray.rllib.policy.sample_batch import DEFAULT_POLICY_ID, MultiAgentBatch
-from ray.rllib.utils.test_utils import check, get_cartpole_dataset_reader
+from ray.rllib.utils.test_utils import get_cartpole_dataset_reader
 
 tf1, tf, tfv = try_import_tf()
 tf1.executing_eagerly()
+
+
+def add_module_helper(env, module_id, runner_or_trainer):
+    runner_or_trainer.add_module(
+        module_id=module_id,
+        module_cls=DiscreteBCTFModule,
+        module_kwargs={
+            "observation_space": env.observation_space,
+            "action_space": env.action_space,
+            "model_config": {"hidden_dim": 32},
+        },
+        optimizer_cls=tf.keras.optimizers.Adam,
+    )
 
 
 class TestTrainerRunner(unittest.TestCase):
@@ -94,7 +107,7 @@ class TestTrainerRunner(unittest.TestCase):
         new_module_id = "test_module"
 
         # add a test_module
-        self.add_module_helper(env, new_module_id, runner)
+        add_module_helper(env, new_module_id, runner)
 
         # do training that includes the test_module
         results = runner.update(
@@ -138,67 +151,6 @@ class TestTrainerRunner(unittest.TestCase):
             self.assertEqual(
                 set(result["loss"]) - {"total_loss"}, module_ids_before_add
             )
-
-    def test_trainer_runner_no_gpus(self):
-        import os
-
-        os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
-        env = gym.make("CartPole-v1")
-        trainer_class = BCTfRLTrainer
-        trainer_cfg = dict(
-            module_class=DiscreteBCTFModule,
-            module_kwargs={
-                "observation_space": env.observation_space,
-                "action_space": env.action_space,
-                "model_config": {"hidden_dim": 32},
-            },
-            optimizer_config={"lr": 1e-3},
-            in_test=True,
-        )
-        runner = TrainerRunner(
-            trainer_class, trainer_cfg, compute_config=dict(num_gpus=0)
-        )
-
-        local_trainer = trainer_class(**trainer_cfg)
-        local_trainer.build()
-
-        # make the state of the trainer and the local runner identical
-        local_trainer.set_state(runner.get_state()[0])
-
-        reader = get_cartpole_dataset_reader(batch_size=500)
-        batch = reader.next()
-        batch = batch.as_multi_agent()
-        check(local_trainer.update(batch), runner.update(batch)[0])
-
-        new_module_id = "test_module"
-
-        # add a test_module
-        self.add_module_helper(env, new_module_id, runner)
-        self.add_module_helper(env, new_module_id, local_trainer)
-
-        # make the state of the trainer and the local runner identical
-        local_trainer.set_state(runner.get_state()[0])
-
-        # do another update
-        batch = reader.next()
-        ma_batch = MultiAgentBatch(
-            {new_module_id: batch, DEFAULT_POLICY_ID: batch}, env_steps=batch.count
-        )
-        check(local_trainer.update(ma_batch), runner.update(ma_batch)[0])
-
-        check(local_trainer.get_state(), runner.get_state()[0])
-
-    def add_module_helper(self, env, module_id, runner_or_trainer):
-        runner_or_trainer.add_module(
-            module_id=module_id,
-            module_cls=DiscreteBCTFModule,
-            module_kwargs={
-                "observation_space": env.observation_space,
-                "action_space": env.action_space,
-                "model_config": {"hidden_dim": 32},
-            },
-            optimizer_cls=tf.keras.optimizers.Adam,
-        )
 
 
 if __name__ == "__main__":
